+++ ttnn/cpp/ttnn/operations/experimental/ccl/ring_attention_all_gather_async/device/kernels/ring_attention_all_gather_writer.cpp
@@ -154,7 +154,7 @@ void kernel_main() {
                             pkt_hdr,
                             *fabric_direction_connection,
                             l1_read_addr,
-                            output_page_size,
+                            (uint16_t)output_page_size,
                             output_page_size);
                     }
 
@@ -295,7 +295,7 @@ void kernel_main() {
                             pkt_hdr,
                             *fabric_direction_connection,
                             l1_read_addr,
-                            output_page_size,
+                            (uint16_t)output_page_size,
                             output_page_size);
                     } else {
                         ASSERT(num_pages_to_read == 1);
